+++ src/test/java/LoginTests.java
@@ -1,83 +1,31 @@
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pom.LoginPage;
-import pom.BasePage;
 import pom.HomePage;
+import pom.LoginPage;
 
-public class LoginTests extends BaseTest{
-    String url = "https://qa.koel.app/";
-    @Test
-    public void loginEmptyEmailTest() {
-        // GIVEN
 
+public class LoginTests extends BaseTest {
+
+
+    @Test(dataProvider = "InvalidLoginCredentials",dataProviderClass = BaseTest.class,enabled = true,priority = 0, description = "Login with invalid email and password")
+    public void LoginInvalidCredentialsTest(String email,String password){
+
+//       driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
         LoginPage loginPage = new LoginPage(getDriver());
-
-        // WHEN
-        loginPage.provideEmail("")
-                .providePassword("te$t$tudent")
-                .clickSubmitBtn();
-
+        String url = "https://qa.koel.app/";
+        loginPage.provideEmail(email).providePassword(password).clickSubmitBtn();
         Assert.assertEquals(getDriver().getCurrentUrl(),url);
+
     }
 
-    @Test
-    public void loginWrongPasswordTest() {
-        // GIVEN
-        LoginPage loginPage = new LoginPage(getDriver());
-
-        // WHEN
-        loginPage.provideEmail("demo@class.com")
-                .providePassword("te$t123")
-                .clickSubmitBtn();
-
-        // THEN
-        Assert.assertEquals(getDriver().getCurrentUrl(),url);
-    }
-
-    @Test
-    public void loginEmptyPasswordTest() {
-        // GIVEN
-        LoginPage loginPage = new LoginPage(getDriver());
-
-        // WHEN
-        loginPage.provideEmail("demo@class.com")
-                .providePassword("")
-                .clickSubmitBtn();
-
-        // THEN
-        Assert.assertEquals(getDriver().getCurrentUrl(),url);
-    }
-
-    @Test
-    public void loginWrongEmailTest() {
-        // GIVEN
-        LoginPage loginPage = new LoginPage(getDriver());
-
-        // WHEN
-        loginPage.provideEmail("wrongemail@class.com")
-                .providePassword("te$t$tudent")
-                .clickSubmitBtn();
-
-        // THEN
-        Assert.assertEquals(getDriver().getCurrentUrl(),url);
-    }
-
-    @Test
-    public void loginValidEmailPasswordTest() {
-        // GIVEN
+    @Test(enabled = true,priority = 1,description = "Login with valid email and valid password")
+    public  void LoginValidEmailPasswordTest() {
         LoginPage loginPage = new LoginPage(getDriver());
         HomePage homePage = new HomePage(getDriver());
-
-        // WHEN
-        loginPage.provideEmail("demo@class.com");
-        loginPage.providePassword("te$t$tudent");
-        loginPage.clickSubmitBtn();
-
-        // THEN
+        //     loginPage.login();
+        loginPage.provideEmail("faiz.ousmail@testpro.io").providePassword("te$t$tudent1").clickSubmitBtn();
         Assert.assertTrue(homePage.getUserAvatar().isDisplayed());
     }
+
+
 }
-
-
-
-
