+++ src/test/resources/Smoke.xml
@@ -6,7 +6,7 @@
     <test name="Example Test" preserve-order="false">
         <classes>
             <class name="BaseTest"/>
-            <class name="LoginTestsDataProvider">
+            <class name="LoginTests">
               <methods>
                   <include name="renamePlaylist"/>
               </methods>
