+++ src/test/resources/TestNG.xml
@@ -6,7 +6,7 @@
     <parameter name="BaseURL" value="https://qa.koel.app/"/>
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="LoginTestsDataProvider"/>
+            <class name="LoginTests"/>
             <class name="BaseTest"/>
 <!--            <class name="Homework16"/>-->
 <!--            <class name="Homework17"/>-->
